Connect SoundController with SoundPool
