+++ Assets/Academy-Platformer/GameController/GameController.cs
@@ -1,3 +1,4 @@
+using Academy_Platformer.Sounds;
 using FactoryPlayer;
 
 public class GameController
@@ -5,6 +6,7 @@ public class GameController
     private FallObjectSpawner _spawner;
 
     private PlayerController _playerController;
+    private SoundController _soundController;
 
     public GameController(
         FallObjectSpawner spawner,
@@ -13,6 +15,8 @@ public class GameController
     {
         _spawner = spawner;
         _playerController = playerController;
+
+        _soundController = new SoundController();
     }
 
     public void StartGame()
@@ -27,7 +31,7 @@ public class GameController
         _spawner.Pool.AllReturnToPool();
         TickableManager.UpdateNotify -= Update;
     }
-    
+
     void Update()
     {
         _spawner.Update();
