+++ Assets/Academy-Platformer/ResourcesConst.cs
@@ -8,10 +8,8 @@
         public const string PlayerConfig = "PlayerConfig";
         public const string TickableManager = "TickableManager";
         public const string MainCamera = "Main Camera";
-        public const string SoundManager = "SoundManager";
         public const string SoundConfig = "SoundConfig";
-        public const string ScoreCounter = "Score Counter";
-        public const string SoundPoolView = "SoundPoolView";
         public const string SoundView = "SoundView";
+        public const string ScoreCounter = "Score Counter";
     }
 }
