--- Assets/Academy-Platformer/Sound.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f02e05e29508b7c4dbb3ff46461b7780
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
