--- Assets/Academy-Platformer/SoundPool/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 90880ee88e390444983da0d1cd868ac8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
