--- Assets/Academy-Platformer/SoundPool/Resources/SoundConfig.asset
@@ -1,35 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 644ae195c49c497a92eb30a5283be2b1, type: 3}
-  m_Name: SoundConfig
-  m_EditorClassIdentifier: 
-  soundModel:
-  - NameSound: BackMain
-    Sound: {fileID: 8300000, guid: f49157a9dd965a64089f52a456a35d4f, type: 3}
-  - NameSound: BackStart
-    Sound: {fileID: 8300000, guid: c9e5e6de17c6edd47880c02024de35a3, type: 3}
-  - NameSound: Buff1
-    Sound: {fileID: 8300000, guid: 0446b457875e6f140ba51f166babd4e9, type: 3}
-  - NameSound: Buff2
-    Sound: {fileID: 8300000, guid: 1b1a5469dfc007b4ea1175c9ba8247d8, type: 3}
-  - NameSound: Debuff1
-    Sound: {fileID: 8300000, guid: 0a13117bbf47bfe4f883f81dff4c9b15, type: 3}
-  - NameSound: Debuff2
-    Sound: {fileID: 8300000, guid: f755659d7beb1ad428f3feecdaa89262, type: 3}
-  - NameSound: GameOver
-    Sound: {fileID: 8300000, guid: 941ab91a14f78414a9c089559ebbb628, type: 3}
-  - NameSound: GetDamage
-    Sound: {fileID: 8300000, guid: d72b48b392bc4364ca83379bdc489081, type: 3}
-  - NameSound: Restart
-    Sound: {fileID: 8300000, guid: 593e70f7b40d34f489f1c6ddcac143d8, type: 3}
-  - NameSound: StartGame
-    Sound: {fileID: 8300000, guid: 2c5dd91a54022dc4fb68f9b4cfa18b8e, type: 3}
