--- Assets/Academy-Platformer/SoundPool/Resources/SoundConfig.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d39ffa6bec0ff5849865394b259a2fb5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
