--- Assets/Academy-Platformer/SoundPool/Resources/SoundPoolView.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8a81eeb223b5f1a4e9182cb533233672
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
