--- Assets/Academy-Platformer/SoundPool/SoundConfig.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 644ae195c49c497a92eb30a5283be2b1
-timeCreated: 1692251028
