--- Assets/Academy-Platformer/SoundPool/SoundPoolView.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-
-namespace SoundPool
-{
-    public class SoundPoolView : MonoBehaviour
-    {
-        
-    }
-}
