--- Assets/Academy-Platformer/SoundPool/SoundPoolView.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 607a3e4a99144b768ab400fe35fe6ae6
-timeCreated: 1692340802
