--- Assets/Academy-Platformer/SoundPool/SoundView.cs
@@ -1,11 +1,0 @@
-﻿using UnityEngine;
-
-namespace SoundPool
-{
-    public class SoundView : MonoBehaviour
-    {
-        public AudioSource AudioSource => audioSource;
-
-        [SerializeField] private AudioSource audioSource;
-    }
-}
