--- Assets/Academy-Platformer/SoundPool/SoundView.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 5192263433374aa1982ecfe7a2db7ad9
-timeCreated: 1692340053
