+++ Assets/Academy-Platformer/Sounds/Resources/SoundConfig.asset
@@ -16,20 +16,20 @@ MonoBehaviour:
   - Name: 0
     Clip: {fileID: 8300000, guid: 3a0c3e40eb6b98c48bdd9e17a8fd427c, type: 3}
   - Name: 1
-    Clip: {fileID: 8300000, guid: 514a75b694518cd4d9350626a9a0ccfa, type: 3}
+    Clip: {fileID: 8300000, guid: 759f87b43e36bb34d9ce6b851c6735c9, type: 3}
   - Name: 2
-    Clip: {fileID: 8300000, guid: 3c41bbff03a6c91488c6f1d14a5951f8, type: 3}
+    Clip: {fileID: 8300000, guid: ef598c6f4840e50418467f48a6072ebb, type: 3}
   - Name: 3
-    Clip: {fileID: 8300000, guid: 9b1a37593affde949806be1331b4010b, type: 3}
+    Clip: {fileID: 8300000, guid: ee6b18e548540d241b1261cad571f59e, type: 3}
   - Name: 4
-    Clip: {fileID: 8300000, guid: 6fdc583ecfd78c04596e1ccd8c78bb50, type: 3}
+    Clip: {fileID: 8300000, guid: ab600e8a7eaf40b44b511975f234e6b2, type: 3}
   - Name: 5
-    Clip: {fileID: 8300000, guid: 1301fa06259d9a54ebdf41c91ae43d86, type: 3}
+    Clip: {fileID: 8300000, guid: f5e4859b3e94f2f449f9ed9325ed7c8d, type: 3}
   - Name: 6
-    Clip: {fileID: 8300000, guid: 3d5e5ab3da569c445bc7a2cad9f2aa78, type: 3}
+    Clip: {fileID: 8300000, guid: 111007146943fbb468e9cb876ac5598c, type: 3}
   - Name: 7
-    Clip: {fileID: 8300000, guid: 1e347b12ee1b28b4aa345ce5dcaa96fd, type: 3}
+    Clip: {fileID: 8300000, guid: 8763b887a50151444ad98f4151b0e12c, type: 3}
   - Name: 8
-    Clip: {fileID: 8300000, guid: 755e1482ecc0f1140b0285f5bd7041d2, type: 3}
+    Clip: {fileID: 8300000, guid: e1606586182972246baf82367acb5ba1, type: 3}
   - Name: 9
-    Clip: {fileID: 8300000, guid: d93a78daa5806f64790294c50a7ae5f8, type: 3}
+    Clip: {fileID: 8300000, guid: 8a893095d2b2ce740960ba8732604113, type: 3}
