--- Assets/Academy-Platformer/Sounds/Resources/SoundManager.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 90aee6f9d5df4be429a0c271dec1f624
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
