--- Assets/Academy-Platformer/Sounds/Resources/SoundsList/SatrtGame.mp3.meta
@@ -1,22 +1,0 @@
-fileFormatVersion: 2
-guid: 0b2e7497c0390cb4f870d05c1e8181b5
-AudioImporter:
-  externalObjects: {}
-  serializedVersion: 6
-  defaultSettings:
-    loadType: 0
-    sampleRateSetting: 0
-    sampleRateOverride: 44100
-    compressionFormat: 1
-    quality: 1
-    conversionMode: 0
-  platformSettingOverrides: {}
-  forceToMono: 0
-  normalize: 1
-  preloadAudioData: 1
-  loadInBackground: 0
-  ambisonic: 0
-  3D: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
