+++ Assets/Academy-Platformer/Sounds/SoundPool.cs.meta
