+++ Assets/Academy-Platformer/Sounds/SoundView.cs
@@ -1,7 +1,11 @@
-using UnityEngine;
+﻿using UnityEngine;
 
-namespace Academy_Platformer.SoundMVC
+namespace Academy_Platformer.Sounds
 {
     public class SoundView : MonoBehaviour
-    { }
+    {
+        public AudioSource AudioSource => audioSource;
+
+        [SerializeField] private AudioSource audioSource;
+    }
 }
