+++ Assets/Academy-Platformer/Sounds/SoundView.cs.meta
@@ -1,11 +1,3 @@
-fileFormatVersion: 2
-guid: 4f18fad5baac6134c91bf52241c1f193
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: 5192263433374aa1982ecfe7a2db7ad9
+timeCreated: 1692340053
